Drop Coercer in favor of `option?` boolean methods
This is the only coercion we actually want to
specify. For anything else, Ruby has fine behavior;
e.g., `to_i`. We just want to treat things like
'0' from a form as false.

We also need to leave the accessors alone, so that
they return the raw option value, so we can use
that value in "sticky" forms; you run a search,
the page reloads, and the search form still contains
what you typed in, not a coerced value.

The nicest way to accomplish both is to just define
separate boolean accessors. `foo` gets you the raw
foo option and `foo?` coerces to a boolean.