Massive rewrite to remove DSL - YOLO! :D
OK, it’s not really *that* massive, because this is a smallish gem.

After a hiatus from using or working on Searchlight, I started to use it
in a project and found it a bit confusing. “What’s this about
`search_on` and procs and procs that call procs?”, I asked myself. “Why
isn’t that just an instance method?"

I was annoyed that the tool I wrote for building searches was not the
tool I wanted to use. Close, but not quite. So I fixed it.

These are breaking changes, but adjusting to the new way of doing things
is pretty simple. And I’ve worked on an example app in parallel with
this code, in addition to writing tests, to ensure that the new code is
nice to use. And it is.

Enjoy!